+++ scripts/dialogs/preview.mjs
@@ -1,35 +1,33 @@
+// GET REQUIRED LIBRARIES
+import '../libraries/marked.min.js';
+import '../libraries/purify.min.js';
+
 // GET MODULE CORE
 import { MODULE } from '../_module.mjs';
 
 export class PreviewDialog extends FormApplication {
-	constructor(data, type = 'changelog') {
+	constructor(data) {
 		super();
 
-		// Sort Items by User Friendly Title
-		const sortedItems = Object.entries(data).sort((a, b) => {
-			return (a[1]?.title.toUpperCase() ?? "UNKNOWN MODULE TITLE") > (b[1]?.title.toUpperCase() ?? "UNKNOWN MODULE TITLE") ? 1 : -1
-		}).reduce((obj, value) => {
-			obj[value[0]] = value[1]
-			return obj;
-		}, {});
-
-		this.items = sortedItems;
-		this.type = type;
+		this.items = data;
+		this.IsSystem = Object.values(data)[0]?.isSystem ?? false;
 	}
-
+  
 	static get defaultOptions() {
-		return {
-			...super.defaultOptions,
+		return mergeObject(super.defaultOptions, {
 			title: MODULE.TITLE,
 			id: `${MODULE.ID}-dialog`,
-			template: `./modules/${MODULE.ID}/templates/preview.hbs`,
+			classes: ['sheet', 'journal-sheet', 'journal-entry'],
+			popOut: true,
 			resizable: true,
-			width: $(window).width() > 720 ? 720 : $(window).width() - 100,
-			height: $(window).height() > 600 ? 600 : $(window).height() - 100
-		}
+			template: `./modules/${MODULE.ID}/templates/preview.hbs`,
+			width: $(window).width() > 960 ? 960 : $(window).width() - 100,
+			height: $(window).height() > 800 ? 800 : $(window).height() - 100
+		});
 	}
-
+  
 	getData() {
+		// Send data to the template
 		return {
 			DIALOG: {
 				ID: MODULE.ID,
@@ -38,38 +36,138 @@ export class PreviewDialog extends FormApplication {
 			items: this.items,
 			IsMultipleItems: Object.keys(this.items).length > 1,
 			IsGM: game.user.isGM
-		}
+		};
 	}
-
+  
 	activateListeners(html) {
 		super.activateListeners(html);
-		
-		$(html).on('click', `nav ul li a.${MODULE.ID}-dialog-link`, (event) => {
-			let $element = $(event.target);
-			const moduleID = $element.data('load');
 
-			// Set as active
-			$(html).find('nav ul li.active').removeClass('active');
-			$(event.target).closest('li').addClass('active');
+		// Get Changelogs and Bind Events
+		const navElements = html[0].querySelectorAll('aside.sidebar nav ol.directory-list li.directory-item');
+		navElements.forEach((navElem) => {
+			navElem.querySelectorAll('span').forEach(spanElem => {
+				spanElem.addEventListener('click', (event) => {
+					const elem = event.target.closest('li');
 
-			game.modules.get(MODULE.ID).API.getContent(moduleID, this.type).then(response => {
-				$(html).find(`.${MODULE.ID}-dialog-title span`).html($element.html());
-				$(html).find(`.${MODULE.ID}-dialog-content`).html(MODULE.markup(response.replaceAll('](/', ']('), { baseUrl: 'modules/monks-tokenbar/' }));
+					// Uncheck active changelog
+					navElements.forEach(activeElem => activeElem.classList.remove('active'));
+					// Check new changelog as active
+					elem.classList.add('active');
+					// Set changelog content
+					game.modules.get(MODULE.ID).API.getContent(this.IsSystem ? game.system : game.modules.get(elem.dataset.moduleId), elem.dataset.fileType, { dir: this.IsSystem ? 'systems' : 'modules'}).then(response => {
+						let content = DOMPurify.sanitize(marked.parse(response, {}), mergeObject({USE_PROFILES: {html: true}}, {}, { inplace: false }));
+						html[0].querySelector('section.journal-entry-content .journal-header input.title').value = (this.IsSystem ? game.system : game.modules.get(elem.dataset.moduleId)).title;
+						html[0].querySelector('section.journal-entry-content .journal-entry-page').innerHTML = content;
 
-				// Update Has Seen Status
-				if (game.user.isGM && this.type == 'changelog') {
-					$(event.target).closest('li').removeClass(`${MODULE.ID}-has-seen-false`);
-					$(event.target).closest('li').addClass(`${MODULE.ID}-has-seen-true`);
-					MODULE.setting('trackedChangelogs', mergeObject(MODULE.setting('trackedChangelogs'), { [moduleID]: { hasSeen: true } }));
-				}
-			})
+						// Update Module as Seen
+						elem.querySelector('.page-ownership i.fa-regular').classList.remove('fa-eye-slash');
+						elem.querySelector('.page-ownership i.fa-regular').classList.add('fa-eye');
+						elem.querySelector('.page-ownership').classList.add('observer');
+						if (game.user.isGM && (MODULE.setting('trackedChangelogs')[elem.dataset.moduleId] ?? false)) {
+							MODULE.setting('trackedChangelogs', foundry.utils.mergeObject(MODULE.setting('trackedChangelogs'), { [elem.dataset.moduleId]: { hasSeen: true } }));
+						}
+					}).catch(result => {
+						MODULE.error(result);
+					});
+
+					// Disable Buttons
+					html[0].querySelectorAll('aside.sidebar .action-buttons button').forEach(btn => btn.disabled = false);
+					if (navElements[0].classList.contains('active')) html[0].querySelector('aside.sidebar .action-buttons button[data-action="previous"]').disabled = true;
+					if (navElements[navElements.length - 1].classList.contains('active')) html[0].querySelector('aside.sidebar .action-buttons button[data-action="next"]').disabled = true;
+				});
+			});
 		});
 
-		$(html).find(`nav ul li:first-child a.${MODULE.ID}-dialog-link`).trigger('click');
-
-
-		$(html).find(`.${MODULE.ID}-dialog-toggle`).on('click', (event) => {
-			$(html).find(`.${MODULE.ID}-window`).toggleClass('full-width')
+		// Bind Collapse Toggle
+		const collapseElem = html[0].querySelector('aside.sidebar .directory-header a.action-button.collapse-toggle');
+		collapseElem.addEventListener('click', (event) => {
+			const app = html[0].closest('.app');
+			const sidebar = app.querySelector(".sidebar");
+			const button = sidebar.querySelector(".collapse-toggle");
+			const sidebarCollapsed = !sidebar.classList.contains('collapsed');
+		
+			// Disable application interaction temporarily
+			app.style.pointerEvents = "none";
+		
+			// Configure CSS transitions for the application window
+			app.classList.add("collapsing");
+			app.addEventListener("transitionend", () => {
+				app.style.pointerEvents = "";
+				app.classList.remove("collapsing");
+			}, {once: true});
+		
+			// Learn the configure sidebar widths
+			const style = getComputedStyle(sidebar);
+			const expandedWidth = Number(style.getPropertyValue("--sidebar-width-expanded").trim().replace("px", ""));
+			const collapsedWidth = Number(style.getPropertyValue("--sidebar-width-collapsed").trim().replace("px", ""));
+		
+			// Change application position
+			const delta = expandedWidth - collapsedWidth;
+			this.setPosition({
+			  left: this.position.left + (sidebarCollapsed ? delta : -delta),
+			  width: this.position.width + (sidebarCollapsed ? -delta : delta)
+			});
+		
+			// Toggle display of the sidebar
+			sidebar.classList.toggle("collapsed", sidebarCollapsed);
+		
+			// Update icons and labels
+			button.dataset.tooltip = sidebarCollapsed ? "JOURNAL.ViewExpand" : "JOURNAL.ViewCollapse";
+			const i = button.children[0];
+			i.setAttribute("class", `fa-solid ${sidebarCollapsed ? "fa-caret-left" : "fa-caret-right"}`);
+			game.tooltip.deactivate();
 		})
+
+		// Bind Search Input
+		const searchInput = html[0].querySelector('aside.sidebar .directory-header input[type="text"]');
+		searchInput.addEventListener('keyup', (event) => {
+			const searchFor = event.target.value.toLowerCase();
+
+			navElements.forEach(navElem => {
+				navElem.classList.remove('hidden');
+
+				if (searchFor == "") {
+					navElem.classList.remove('hidden');
+				}else if (navElem.querySelector('.page-title').textContent.toLowerCase().indexOf(searchFor) == -1 && !navElem.classList.contains('active')) {
+					navElem.classList.add('hidden');
+				}
+			})
+		})
+
+		// Bind Next/Prev Buttons
+		const nextButton = html[0].querySelector('aside.sidebar .action-buttons button[data-action="next"]');
+		const previousButton = html[0].querySelector('aside.sidebar .action-buttons button[data-action="previous"]');
+
+		nextButton.addEventListener('click', (event) => {
+			html[0].querySelector('aside.sidebar nav ol.directory-list li.directory-item.active').nextElementSibling.querySelector('.page-title').click();
+		});
+		previousButton.addEventListener('click', (event) => {
+			html[0].querySelector('aside.sidebar nav ol.directory-list li.directory-item.active').previousElementSibling.querySelector('.page-title').click();
+		});
+
+		// Mark as Seen
+		const markAsSeen = html[0].querySelector('aside.sidebar .action-buttons button[data-action="markAsSeen"]');
+
+		markAsSeen.addEventListener('click', event => {
+			let trackedChangelogs = MODULE.setting('trackedChangelogs');
+			for (let key in trackedChangelogs) {
+				trackedChangelogs[key].hasSeen = true;
+
+				if (html[0].querySelector(`aside.sidebar nav ol.directory-list li[data-module-id="${key}"]`) ?? false) {
+					html[0].querySelector(`aside.sidebar nav ol.directory-list li[data-module-id="${key}"] span.page-ownership`).classList.add('observer');
+					html[0].querySelector(`aside.sidebar nav ol.directory-list li[data-module-id="${key}"] span.page-ownership i`).classList.remove('fa-eye-slash');
+					html[0].querySelector(`aside.sidebar nav ol.directory-list li[data-module-id="${key}"] span.page-ownership i`).classList.add('fa-eye');
+				}
+			}
+
+			MODULE.setting('trackedChangelogs', trackedChangelogs);
+		})
+
+
+		navElements[0].querySelector('.page-title').click();
 	}
-}
+  
+	async _updateObject(event, formData) {
+		console.log(formData.exampleInput);
+	}
+  }